phox-container/requirements.txt
boto3
xgboost==2.0.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.2.2
pillow==9.4.0